Move plaato coordinator to separate module
